feat(script/sortArchive): add sorted indicator
As a reaction to releasing Ctrl key right before selecting items in
archive, hopefully this is a good enough of a reminder.

A better solution would block interaction right after the key has been
released, but I am not 100% confident that there won't be an interaction
that causes the webpage to get stuck in that state, making the game harder to play.